tools/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "OldSchoolTibia-tools"
version = "0.6"
dependencies = [
    "pycryptodome==3.15.0"
]
authors = [
    {name = "Simon Sandström"}
]
license = "MIT"

[project.urls]
Homapage = "https://github.com/gurka/OldSchoolTibia"